+++ src/Server/Process/Success.php
@@ -3,7 +3,7 @@ declare(strict_types = 1);
 
 namespace Innmind\Server\Control\Server\Process;
 
-final class Signaled
+final class Success
 {
     private Output $output;